Parse string values as double not integer

+++ Shared/Vector.java
@@ -17,8 +17,8 @@ public class Vector {
 
     public Vector(String stringValue) {
         String[] split = stringValue.split(" ");
-        x = Integer.valueOf(split[0]);
-        y = Integer.valueOf(split[1]);
+        x = Double.valueOf(split[0]);
+        y = Double.valueOf(split[1]);
     }
 
     public double getX() {
